created util class
