+++ public/NewSTAdvancedFunction.cs
@@ -16,9 +16,7 @@ namespace SimpleTemplates
             ValueFromPipelineByPropertyName = true)]
         public string[] FunctionName { get; set; }
         private const string  TemplateFileName = "templates/AdvancedFunction.ps1t";
-        private string CurrentDirectory;
 
-        // This method gets called once for each cmdlet in the pipeline when the pipeline starts executing
         protected override void BeginProcessing()
         {
             // check if template file exists
@@ -26,25 +24,12 @@ namespace SimpleTemplates
                 throw new FileNotFoundException("Template file not found", TemplateFileName);
             }
             WriteVerbose($"Using template file {TemplateFileName}");
-            CurrentDirectory = GetUnresolvedProviderPathFromPSPath(this.SessionState.Path.CurrentFileSystemLocation.ToString());
         }
 
-        // This method will be called for each input received from the pipeline to this cmdlet; if no input is received, this method is not called
         protected override void ProcessRecord()
         {
-            string FileName;
             foreach (string F1 in FunctionName) {
-                FileName = Path.Combine(CurrentDirectory, $"{F1}.ps1");
-                if (File.Exists(FileName)) {
-                    WriteWarning($"File {FileName} exists");
-                    continue;
-                }
-                WriteVerbose($"Creating file {FileName}...");
-                File.Create(FileName).Dispose();
-                WriteVerbose($"Created {FileName}");
-                foreach (string line in File.ReadLines(TemplateFileName)) {
-                    File.AppendAllText(FileName, line.Replace("$1", F1) + Environment.NewLine);
-                }
+                Util.CreateFileFromTemplate(this, TemplateFileName, F1);
             }
         }
